refactor and error handling
