+++ libs/dh/charges/feature-prices/src/lib/drawer/charge-content/details-header/dh-charge-details-header.component.scss
