+++ libs/dh/charges/feature-prices/src/lib/drawer/charge-message/dh-charge-price-message.component.html
@@ -18,17 +18,17 @@ limitations under the License.
   <watt-button (click)="backToCharge()" [variant]="'text'" [icon]="'left'">
     {{ t("back") }}</watt-button
   >
-  <br />
-  <small>{{ message?.createdDate | dhDateTime }}</small>
-  <h1>{{ message?.messageId }}</h1>
-  <small>
-    {{ "charges.messageType." + message?.processType | transloco }} |
-    {{ "charges.messageTypeAbbreviation." + message?.processType | transloco }}
-  </small>
+  <ng-container *ngIf="message">
+    <br />
+    <small>{{ message.createdDate | dhDateTime }}</small>
+    <h1>{{ message.messageId }}</h1>
+    <small>
+      {{ "charges.messageType." + message.processType | transloco }} |
+      {{ "charges.messageTypeAbbreviation." + message.processType | transloco }}
+    </small>
 
-  <mat-divider></mat-divider>
+    <mat-divider></mat-divider>
 
-  <div class="test">
     <label>{{ t("sender") }}</label>
     <small>
       {{ senderMarketParticipant?.name }} ({{
@@ -45,16 +45,26 @@ limitations under the License.
     >
       {{ t("prices.Download") }}
     </watt-button>
-  </div>
+  </ng-container>
 
   <!-- Show message -->
   <ng-container *rxLet="blobContent$ as blobContent">
     <pre>{{ blobContent }}</pre>
   </ng-container>
 
-  <div *ngIf="blobIsDownloading$ | push">
+  <div *ngIf="(blobIsDownloading$ | push) && (messageIsSearching$ | push)">
     <watt-spinner [diameter]="60"></watt-spinner>
   </div>
+
+  <!-- Error -->
+  <ng-container *ngIf="messageHasGeneralError$ | push">
+    <watt-empty-state
+      class="loading-error"
+      icon="power"
+      title="{{ t('prices.loadingErrorTitle') }}"
+      message="{{ t('prices.loadingErrorMessage') }}"
+    ></watt-empty-state>
+  </ng-container>
 
   <div *transloco="let transloco; read: 'messageArchive'">
     <watt-badge *ngIf="blobHasGeneralError$ | push" type="danger">{{
